+++ src/__tests__/useMutation-test.js
@@ -5,7 +5,6 @@ import {
   waitForElement,
 } from '@testing-library/react'
 import * as React from 'react'
-import { act } from 'react-dom/test-utils'
 
 import { useMutation } from '../index'
 
@@ -42,5 +41,40 @@ describe('useMutation', () => {
     await waitForElement(() => getByTestId('title'))
 
     expect(getByTestId('title').textContent).toBe('')
+  })
+
+  it('should be able to reset `error`', async () => {
+    function Page() {
+      const [mutate, mutationResult] = useMutation(
+        () => Promise.reject(new Error('something went wrong')),
+        {
+          throwOnError: false
+        }
+      )
+
+      return (
+        <div>
+          {mutationResult.error &&
+            <h1 data-testid="error">{mutationResult.error.message}</h1>
+          }
+          <button onClick={mutationResult.reset}>reset</button>
+          <button onClick={mutate}>mutate</button>
+        </div>
+      )
+    }
+
+    const { getByTestId, getByText, queryByTestId } = render(<Page />)
+
+    expect(queryByTestId('error')).toBeNull()
+
+    fireEvent.click(getByText('mutate'))
+
+    await waitForElement(() => getByTestId('error'))
+
+    expect(getByTestId('error').textContent).toBe('something went wrong')
+
+    fireEvent.click(getByText('reset'))
+
+    expect(queryByTestId('error')).toBeNull()
   })
 })